+++ src/extensions/admin/account/devices/UserDeviceTablePage.js
@@ -32,7 +32,7 @@ class UserDeviceTablePage extends TablePage {
                 { key: 'id', label: 'ID', width: '70px', sortable: true, class: 'text-muted' },
                 { key: 'duid', label: 'Device ID', sortable: true, formatter: 'truncate_middle(16)', visibility: 'xl' },
                 { key: 'user.display_name', label: 'User', sortable: true, formatter: "default('—')" },
-                { key: 'device_info.user_agent.family', label: 'Browser', formatter: "default('—')" },
+                { key: 'device_info.user_agent.family', label: 'Browser', formatter: "default('—')", visibility: 'md' },
                 { key: 'device_info.os.family', label: 'OS', formatter: "default('—')", visibility: 'lg' },
                 { key: 'last_ip', label: 'Last IP', sortable: true },
                 { key: 'first_seen', label: 'First Seen', formatter: "epoch|datetime", visibility: 'lg' },