+++ src/extensions/admin/messaging/email/EmailMailboxTablePage.js
@@ -24,7 +24,7 @@ class EmailMailboxTablePage extends TablePage {
             columns: [
                 { key: 'id', label: 'ID', width: '70px', sortable: true, class: 'text-muted' },
                 { key: 'email', label: 'Email', sortable: true },
-                { key: 'domain.name', label: 'Domain', sortable: true, formatter: "default('—')" },
+                { key: 'domain.name', label: 'Domain', sortable: true, formatter: "default('—')", visibility: 'md' },
                 { key: 'allow_inbound', label: 'Inbound', formatter: "boolean|badge" },
                 { key: 'allow_outbound', label: 'Outbound', formatter: "boolean|badge" },
                 { key: 'is_system_default', label: 'System Default', formatter: "boolean|badge", visibility: 'lg' },