+++ src/extensions/admin/messaging/push/PushDeliveryTablePage.js
@@ -30,7 +30,7 @@ class PushDeliveryTablePage extends TablePage {
             columns: [
                 { key: 'id', label: 'ID', width: '70px' },
                 { key: 'created', label: 'Timestamp', formatter: 'datetime' },
-                { key: 'user.display_name', label: 'User' },
+                { key: 'user.display_name', label: 'User', visibility: 'md' },
                 { key: 'device.device_name', label: 'Device', visibility: 'md' },
                 { key: 'title', label: 'Title' },
                 { key: 'category', label: 'Category', visibility: 'lg' },